# Further improve the media attached status query for accounts