@ClearlyClaire 2021-04-25
Is there any benefit to this besides re-adding the explicit filtering of media_attachments by account `id`? If not I'm not sure about this, since media attachments only get attached to statuses if they are from the same account anyway.