@abcang 2021-04-25
This change will improve query speed for accounts with few media_attachments or accounts with a low percentage of posts with media_attachments.

It is possible to search more efficiently by using the table with the least amount of data as the driving table. In general, the number of media_attachments in an account is less than the number of statuses, but if media_attachments are not filtered, the statuses table will certainly be the driving table. Therefore, if an account has fewer media_attachments, searching from the statuses table first will be less efficient.

If media_attachments is filtered, the query planner will notice that the media_attachments table has less data and is easier to narrow down. And by adding a composite index, sorting without using indexes will not occur as it happened before.