@ClearlyClaire 2021-04-25
It seems like you forgot to add the db migration script!